requirements.txt
beautifulsoup4==4.8.2
cffi==1.13.2
cryptography==2.8
Django==2.0.13
django-bootstrap4==1.1.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
mysqlclient==1.4.6
pycparser==2.19
PyJWT==1.7.1
PyMySQL==0.9.3
pytz==2019.3
six==1.13.0
soupsieve==1.9.5
